+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Help screen for keybindings
+
+### Changed
+
+- Deprecated Alignment for new HorizontalAlignment
+- Keybinding for switching focus between file explorer and editor
+
+### Removed
+
+
+
+### Fixed
+
+
+
 ## [0.3.2] - 2026-07-23
 
 Initial release